move demo resource to config package
