+++ server/config/demoresource/demo_resource.go
